Simplify dataflow graph setup
